# YJIT: Bounds check every byte in the assembler