@XrXr 2021-12-04
Benchmarks show basically no change to perf:

Before:
```
end_time="2021-12-03 22:25:26 UTC (+0000)"
yjit_opts=""
ruby_version="ruby 3.1.0dev (2021-12-03T17:45:58Z master 54ca530dbe) [x86_64-linux]"
git_branch="overflow"
git_commit="658766efbe"

-------------  -----------  ----------  ---------  ----------  -----------  ------------
bench          interp (ms)  stddev (%)  yjit (ms)  stddev (%)  interp/yjit  yjit 1st itr
30k_ifelse     2351.4       0.1         358.3      0.1         6.56         3.02
30k_methods    6449.2       0.0         881.7      0.0         7.31         6.30
activerecord   162.3        0.4         123.2      0.5         1.32         1.23
binarytrees    446.0        0.1         337.4      0.9         1.32         1.27
cfunc_itself   109.0        0.2         55.4       0.5         1.97         1.99
erubi          476.9        0.6         478.6      0.5         1.00         1.00
erubi_rails    40.5         2.8         34.3       3.0         1.18         1.01
fannkuchredux  6555.4       0.0         6567.1     0.1         1.00         1.00
fib            250.4        0.0         62.7       0.4         3.99         3.99
getivar        109.3        0.7         42.5       0.4         2.57         1.01
hexapdf        3230.5       0.8         2462.4     1.1         1.31         1.23
jekyll         9730.3       1.8         8553.7     2.0         1.14         1.13
keyword_args   286.0        0.0         56.9       0.4         5.02         5.06
lee            1258.2       0.1         901.2      0.8         1.40         1.38
liquid-render  202.9        1.0         133.9      1.6         1.52         1.39
mail           185.5        0.1         166.4      0.2         1.11         0.99
nbody          125.7        0.0         90.2       0.3         1.39         1.38
optcarrot      6195.3       0.7         3697.0     0.6         1.68         1.65
psych-load     2505.8       0.1         2009.2     0.0         1.25         1.24
railsbench     2915.7       0.9         2246.1     1.2         1.30         1.26
respond_to     274.5        0.2         210.4      0.3         1.30         1.29
rubykon        12389.1      0.4         6658.8     0.8         1.86         1.87
setivar        74.6         0.1         49.9       0.8         1.50         1.01
-------------  -----------  ----------  ---------  ----------  -----------  ------------
Legend:
- interp/yjit: ratio of interp/yjit time. Higher is better. Above 1 represents a speedup.
- 1st itr: ratio of interp/yjit time for the first benchmarking iteration.
```

After:
```
end_time="2021-12-04 00:51:30 UTC (+0000)"
yjit_opts=""
ruby_version="ruby 3.1.0dev (2021-12-03T19:35:56Z overflow 658766efbe) [x86_64-linux]"
git_branch="overflow"
git_commit="658766efbe"

-------------  -----------  ----------  ---------  ----------  -----------  ------------
bench          interp (ms)  stddev (%)  yjit (ms)  stddev (%)  interp/yjit  yjit 1st itr
30k_ifelse     2356.8       0.0         358.2      0.2         6.58         2.96
30k_methods    6388.9       0.0         881.6      0.0         7.25         6.23
activerecord   172.8        2.7         117.7      0.5         1.47         1.27
binarytrees    437.1        0.1         341.3      0.2         1.28         1.27
cfunc_itself   103.3        0.3         49.4       0.6         2.09         2.17
erubi          473.5        0.6         480.9      0.7         0.98         0.99
erubi_rails    41.3         2.8         34.1       3.0         1.21         0.84
fannkuchredux  6377.2       0.1         6384.5     0.1         1.00         1.00
fib            252.1        0.0         62.7       0.5         4.02         4.04
getivar        107.0        0.9         42.5       0.6         2.52         1.01
hexapdf        3229.3       0.9         2467.9     1.0         1.31         1.28
jekyll         9776.3       2.4         8567.7     2.0         1.14         1.13
keyword_args   293.4        0.1         57.3       0.5         5.12         5.11
lee            1243.9       0.5         892.7      0.9         1.39         1.39
liquid-render  203.5        0.8         133.9      1.4         1.52         1.40
mail           184.2        0.3         166.4      0.3         1.11         0.98
nbody          123.8        0.0         87.4       0.4         1.42         1.40
optcarrot      6242.4       0.7         3723.8     0.6         1.68         1.65
psych-load     2529.2       0.0         2017.8     0.0         1.25         1.25
railsbench     2917.7       0.9         2248.8     1.2         1.30         1.26
respond_to     275.3        0.3         198.7      0.2         1.39         1.39
rubykon        12227.7      0.4         6545.0     0.3         1.87         1.88
setivar        76.6         1.0         50.4       0.6         1.52         1.01
-------------  -----------  ----------  ---------  ----------  -----------  ------------
Legend:
- interp/yjit: ratio of interp/yjit time. Higher is better. Above 1 represents a speedup.
- 1st itr: ratio of interp/yjit time for the first benchmarking iteration.
```